requirements.txt
flask
flask-sqlalchemy
telethon
werkzeug
gunicorn
python-dotenv
